Move some pytests from lint to utils
